requirements.txt
pydantic
python-dotenv
openai
google-generativeai

